pyproject.toml
[tool.poetry]
name = "pxblat"
version = "0.2.0"
description = "A native python binding for blat suit"
authors = ["Yangyang Li <yangyang.li@northwestern.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.0"
pybind11 = "^2.10.4"
setuptools = ">=67.7.2,<69.0.0"
rich = "^13.3.5"
pysimdjson = "^5.0.2"
biopython = "^1.81"
typer = "^0.9.0"
deprecated = "^1.2.13"
mashumaro = "^3.7"
numpy = "^1.24.3"
gevent = "^22.10.2"


[tool.poetry.scripts]
pxblat = "pxblat.cli.cli:app"


[tool.ruff]
line-length = 120
fix = true
extend-ignore = ["E501", "D203", "D100", "D401"]
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # pylint
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    # misc lints
    "PIE",
    # flake8-pyi
    "PYI",
    # tidy imports
    "TID",
    # implicit string concatenation
    "ISC",
    # type-checking imports
    "TCH",
    # comprehensions
    "C4",
    # flake8-pytest-style
    "PT",
    # "ANN",
    "D",
    "RUF",
]

exclude = [
    "src/pxblat/extc/__init__.pyi",
    "src/pxblat/__init__.py",
    "tasks.py",
    "tests/*",
    "build.py",
    "scripts/*",
    "noxfile.py",
]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.per-file-ignores]
"src/pxblat/extc/__init__.py" = ["F405", "F403", "D104", "D103"]
"src/pxblat/server/basic.py" = ["PLR0913"]
"src/pxblat/server/server.py" = ["PLR0913"]
"src/pxblat/server/client.py" = ["PLR0913"]
"src/pxblat/cli/server.py" = ["PLR0913"]
"src/pxblat/cli/client.py" = ["B008", "PLR0913"]
"src/pxblat/cli/fa2twobit.py" = ["B008", "PLR0913"]
"src/pxblat/cli/twobit2fa.py" = ["B008", "PLR0913"]
"src/pxblat/toolkit/__init__.py" = ["PLR0913"]


[tool.ruff.pydocstyle]
convention = 'google'


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-sugar = "^0.9.7"
black = "^23.3.0"
ruff = ">=0.0.270,<0.0.277"
ipython = "^8.13.1"
pybind11-stubgen = ">=0.13,<0.17"
invoke = "^2.1.2"
ipdb = "^0.13.13"
pysam = "^0.21.0"
pytest-ordering = "^0.6"
pytest-xdist = "^3.3.1"
scienceplots = "^2.0.1"
tikzplotlib = "^0.10.1"
nox = "^2023.4.22"
nox-poetry = "^1.0.2"
codecov = "^2.1.13"
myst-parser = ">=1,<3"
sphinx-immaterial = "^0.11.3"
linkify-it-py = "^2.0.2"
sphinx-copybutton = "^0.5.2"
sphinx-togglebutton = "^0.3.2"
sphinxcontrib-bibtex = "^2.5.0"
sphinx-autoapi = "^2.1.0"
sphinx-autodoc2 = "^0.4.2"
pytest-benchmark = "^4.0.0"


[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["pxblat", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 40
ignore_errors = true


[tool.mypy]
strict = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = ["smoke: a quick smoke test"]


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_type = "pep440"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true

[tool.poetry.build]
script = "build.py"
generate-setup-file = true

[build-system]
requires = ["pybind11>2.9.1", "poetry-core>=1.2.0", "setuptools>=40.6.0"]
build-backend = "poetry.core.masonry.api"
